--- Days/Days/ContentView.swift
@@ -1,21 +1,0 @@
-//
-//  ContentView.swift
-//  Days
-//
-//  Created by Paweł Brzozowski on 26/09/2022.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        Text("Hello, world!")
-            .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
