+++ Monthly/Monthly/Assets.xcassets/AccentColor.colorset/Contents.json
