+++ Monthly/Monthly/Assets.xcassets/Contents.json
