+++ Monthly/Monthly/ContentView.swift
@@ -1,0 +1,26 @@
+//
+//  ContentView.swift
+//  Monthly
+//
+//  Created by Paweł Brzozowski on 30/09/2022.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        VStack {
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundColor(.accentColor)
+            Text("Hello, world!")
+        }
+        .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
