+++ Monthly/Monthly/Preview Content/Preview Assets.xcassets/Contents.json
